`[Config/Queues]` Add RootVolume settings for Queue configuration.
