+++ frontend/src/pages/Configure/Components.js
@@ -35,6 +35,11 @@ import {
 // Components
 import HelpTooltip from '../../components/HelpTooltip'
 
+// Helper Functions
+function strToOption(str){
+  return {value: str, label: str}
+}
+
 const multiRunner = 'https://raw.githubusercontent.com/aws-samples/pcluster-manager/main/resources/scripts/multi-runner.py'
 const knownExtensions = [{name: 'Cloud9', path: 'cloud9.sh', description: 'Cloud9 Install', args: [{name: 'Output File'}]},
   {name: 'Slurm Accounting', path: 'slurm-accounting.sh', description: 'Slurm Accounting', args: [{name: 'Secret ARN'}, {name: 'RDS Endpoint'}, {name: 'RDS Port', default: '3306'}]},
@@ -469,4 +474,62 @@ function SecurityGroups({basePath}) {
   </SpaceBetween>
 }
 
-export { SubnetSelect, SecurityGroups, InstanceSelect, LabeledIcon, ActionsEditor, CustomAMISettings }
+function RootVolume({basePath, errorsPath}) {
+  const rootVolumeSizePath = [...basePath, 'LocalStorage', 'RootVolume', 'Size'];
+  const rootVolumeSize = useState(rootVolumeSizePath);
+
+  const rootVolumeEncryptedPath = [...basePath, 'LocalStorage', 'RootVolume', 'Encrypted'];
+  const rootVolumeEncrypted = useState(rootVolumeEncryptedPath);
+
+  const rootVolumeTypePath = [...basePath, 'LocalStorage', 'RootVolume', 'VolumeType'];
+  const rootVolumeType = useState(rootVolumeTypePath);
+  const volumeTypes = ['gp2', 'gp3', 'io1', 'io2', 'sc1', 'stl', 'standard'];
+
+  const rootVolumeErrors = useState([...errorsPath, 'rootVolume']);
+  const editing = useState(['app', 'wizard', 'editing']);
+
+  const setRootVolume = (size) => {
+    if(size === '')
+      clearState(rootVolumeSizePath);
+    else
+      setState(rootVolumeSizePath, parseInt(size));
+    clearEmptyNest(rootVolumeSizePath, 3);
+  }
+
+  const toggleEncrypted = () => {
+    const setEncrypted = !rootVolumeEncrypted;
+    if(setEncrypted)
+      setState(rootVolumeEncryptedPath, setEncrypted);
+    else
+      clearState(rootVolumeEncryptedPath);
+    clearEmptyNest(rootVolumeSizePath, 3);
+  }
+
+  return <>
+    <FormField
+      label="Root Volume Size (GB)"
+      errorText={rootVolumeErrors}
+      description="Typically users will use a shared storage option for application data so a smaller root volume size is suitable. Blank will use the default from the AMI.">
+      <Input
+        disabled={editing}
+        placeholder="Enter root volume size."
+        value={rootVolumeSize || ''}
+        inputMode="decimal"
+        onChange={({detail}) => setRootVolume(detail.value)} />
+    </FormField>
+    <Toggle
+      disabled={editing}
+      checked={rootVolumeEncrypted || false} onChange={toggleEncrypted}>Encrypted Root Volume</Toggle>
+    <div key="volume-type" style={{display: "flex", flexDirection: "row", alignItems: "center", gap: "16px"}}>
+      Volume Type:
+      <Select
+        disabled={editing}
+        placeholder="Default (gp2)"
+        selectedOption={rootVolumeType && strToOption(rootVolumeType)} label="Volume Type" onChange={({detail}) => {setState(rootVolumeTypePath, detail.selectedOption.value)}}
+        options={volumeTypes.map(strToOption)}
+      />
+    </div>
+  </>
+}
+
+export { SubnetSelect, SecurityGroups, InstanceSelect, LabeledIcon, ActionsEditor, CustomAMISettings, RootVolume }
